+++ src/components/avatar/avatar.vue
@@ -1,19 +1,19 @@
 <template>
   <div :class="avatarClassses">
     <template v-if="!initial">
-      <img v-if="!$slots.default" :src="src" :alt="alt" :class="avatarImageClassses" />
-      <div v-else :class="avatarImageClassses">
+      <img v-if="!$slots.default && src" :src="src" :alt="alt" :class="avatarImageClassses" />
+      <div v-else :class="[avatarImageClassses, 'avatar__slot spr-border-color-weak spr-border spr-border-solid']">
         <slot />
       </div>
     </template>
     <div v-else :class="initialClassses">
       {{ initial }}
     </div>
-    <span v-if="badge" :class="avatarNotificationClassses">
+    <span v-if="notification" :class="avatarNotificationClassses">
       <spr-badge :text="badgeText" variant="danger" :size="getAvatarSize.notif" />
     </span>
-    <span v-if="online" :class="onlineNotificationClassses">
-      <spr-badge text="" variant="brand" :size="getAvatarSize.online" />
+    <span v-if="badge" :class="onlineNotificationClassses">
+      <spr-badge text="" :variant="status" :size="getAvatarSize.badge" />
     </span>
   </div>
 </template>