+++ src/components/empty-state/use-empty-state.ts
@@ -1,24 +1,26 @@
 import { computed, toRefs } from 'vue';
-import type { EmptyStatePropTypes } from './empty-state';
 
 import classNames from 'classnames';
+
+import type { EmptyStatePropTypes } from './empty-state';
+
 export const useEmptyState = (props: EmptyStatePropTypes) => {
   const { size, image } = toRefs(props);
 
   const emptyStateWrapperClasses = computed(() => {
     return classNames(
-      'background-color flex h-full w-full flex-col items-center justify-center gap-size-spacing-2xs px-size-spacing-2xl py-size-spacing-3xl text-center min-h-[240px]',
+      'spr-background-color spr-flex spr-h-full spr-w-full spr-flex-col spr-items-center spr-justify-center spr-gap-size-spacing-2xs spr-px-size-spacing-2xl spr-py-size-spacing-3xl spr-text-center spr-min-h-[240px]',
       {
-        'px-size-spacing-sm py-size-spacing-md min-h-[240px]': size.value === 'small',
-        'px-size-spacing-sm py-size-spacing-2xl min-h-[360px]': size.value === 'large',
+        'spr-px-size-spacing-sm spr-py-size-spacing-md spr-min-h-[240px]': size.value === 'small',
+        'spr-px-size-spacing-sm spr-py-size-spacing-2xl spr-min-h-[360px]': size.value === 'large',
       },
     );
   });
 
   const imageSizeClasses = computed(() => {
     return classNames({
-      'h-[200px] w-[200px]': size.value === 'large',
-      'h-[120px] w-[120px]': size.value === 'small',
+      'spr-h-[200px] spr-w-[200px]': size.value === 'large',
+      'spr-h-[120px] spr-w-[120px]': size.value === 'small',
     });
   });