+++ src/components/modal/use-modal.ts
@@ -1,19 +1,20 @@
 import { ref, watch, computed, toRefs } from 'vue';
-import type { SetupContext } from 'vue';
-import type { ModalEmitTypes } from './modal';
-import type { ModalPropTypes } from './modal';
 
 import classNames from 'classnames';
+
+import type { SetupContext } from 'vue';
+import type { ModalEmitTypes, ModalPropTypes } from './modal';
+
 export const useModal = (props: ModalPropTypes, emit: SetupContext<ModalEmitTypes>['emit']) => {
   const dialog = ref<HTMLDialogElement | null>(null);
   const { size } = toRefs(props);
 
   const modalSizesClasses = computed(() => {
     return classNames({
-      'min-w-[360px] max-w-[800px] ': size.value === 'sm',
-      'min-w-[480px] max-w-[800px] ': size.value === 'md',
-      'w-[720px]': size.value === 'lg',
-      'w-[1280px]': size.value === 'xl',
+      'spr-min-w-[360px] spr-max-w-[800px] ': size.value === 'sm',
+      'spr-min-w-[480px] spr-max-w-[800px] ': size.value === 'md',
+      'spr-w-[720px]': size.value === 'lg',
+      'spr-w-[1280px]': size.value === 'xl',
     });
   });
 
@@ -44,7 +45,6 @@ export const useModal = (props: ModalPropTypes, emit: SetupContext<ModalEmitType
   return {
     dialog,
     modalSizesClasses,
-
     openModal,
     closeModal,
   };