+++ src/components/snackbar/snackbar.vue
@@ -3,12 +3,8 @@
     <Transition name="snackbar">
       <div v-if="snackbarStore.snacks.length > 0" class="snackbar-wrapper slide-in-element">
         <TransitionGroup name="snackbar" tag="ul">
-          <li
-            v-for="snack in snackbarStore.snacks"
-            :key="snack.id"
-            class="snackbar-snack mb-size-spacing-3xs"
-          >
-            <spr-snack 
+          <li v-for="snack in snackbarStore.snacks" :key="snack.id" class="snackbar-snack spr-mb-size-spacing-3xs">
+            <spr-snack
               :text="snack.text"
               :action-text="snack.actionText"
               :tone="snack.tone"
@@ -26,21 +22,23 @@
 
 <script lang="ts" setup>
 import { useSnackbarStore } from '@/stores/useSnackbarStore';
+
 import SprSnack from '@/components/snackbar/snack/snack.vue';
 import { useSnackbar } from '@/components/snackbar/use-snackbar';
+
 import { SnackPropTypes } from './snack/snack';
-import { createPinia } from 'pinia'
+import { createPinia } from 'pinia';
 
 const pinia = createPinia();
 const snackbarStore = useSnackbarStore(pinia);
 const { showSnackbar, showSuccess, showInformation, showDanger, showCaution } = useSnackbar();
 
-defineExpose({ 
-  showSnackbar: (payload: SnackPropTypes) => showSnackbar(payload), 
-  showSuccess: (payload: SnackPropTypes) => showSuccess(payload), 
-  showInformation: (payload: SnackPropTypes) => showInformation(payload), 
-  showDanger: (payload: SnackPropTypes) => showDanger(payload), 
-  showCaution: (payload: SnackPropTypes) => showCaution(payload)
+defineExpose({
+  showSnackbar: (payload: SnackPropTypes) => showSnackbar(payload),
+  showSuccess: (payload: SnackPropTypes) => showSuccess(payload),
+  showInformation: (payload: SnackPropTypes) => showInformation(payload),
+  showDanger: (payload: SnackPropTypes) => showDanger(payload),
+  showCaution: (payload: SnackPropTypes) => showCaution(payload),
 });
 </script>